Version 0.0.5
Final update on 0.0.5 Counter. I did not get it to Display on the Site it self. Maybe i will try at a later date. It is tracked. Then i also cleaned up in the code made it a little more readable and removed some files that is not in use.